line_bot/requirements.txt
Flask==3.0.0
line-bot-sdk==3.6.0
APScheduler==3.10.4
pytz==2023.3
matplotlib==3.8.2
Pillow==10.1.0
google-generativeai==0.3.1
requests==2.31.0
